refactor: 관련 도메인끼리 패키지 분리
