+++ README.md
@@ -25,8 +25,8 @@ Player
 - [x] 카드를 추가할 수 있다.
 - [x] 보유한 카드를 가져올 수 있다. 
 Dealer
-- [] 카드를 추가할 수 있다.
-- [] 보유한 카드를 가져올 수 있다.
+- [x] 카드를 추가할 수 있다.
+- [x] 보유한 카드를 가져올 수 있다.
 
   [게임 결과]
 - [] 딜러와 플레이어간의 승 패 결과
