+++ src/main/java/blackjack/domain/gamer/Person.java
@@ -1,4 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.gamer;
+
+import blackjack.domain.card.Card;
 
 import java.util.ArrayList;
 import java.util.Collections;
